added NEAT + config, MyReporter, DataAccess
Neat and config was implemented allowing for basic logic and configuration of the NEAT AI in Game and GameModels. MyReporter, an extension of Neat.StdOut.Reporter was added to allow the stats of each generation to be accessed. DataAccess, which uses a mysql connector, was added and allows these stats to be posted to the database